updates to be compliant with new fastapi and pydantic
All changes are internal. The main changes are in prototypes and API
errors. Prototypes now are created from dictionaries of tuples of (type, value). API errors are now a list of dictionaries, giving the type, location, error message, erroneous input, and a url to pydantic's website with more about the error.